generalize reactive list
